feature: separate the common features from the client-specific ones
The pusuclt package has been added and a lot of the client-specific parts
have been moved into that so that the pusu package now only contains common
types and code for the server and client.

Also the parameters have been moved into a different module to reduce
dependencies in this module.

The protocol has been changed and the protobuf declarations reflect that.